Refactor customized page, split logic
